feat: update authentication forms and enhance menu styling; add certification images
