+++ front/index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="stylesheet" href="./src/style/index.css">
     <link href="https://fonts.googleapis.com/css2?family=Anton&display=swap" rel="stylesheet">
-    <title>MARVEL FAN | O LUGAR QUE TODOS OS FÃS DESEJAM ESTAR </title>
+    <title>Show Smart | Tenha a sua certificação </title>
   </head>
   <body>
     <div id="root"></div>
